Address  pr comments
