+++ tensorboard/webapp/hparams/_redux/hparams_data_source_test.ts
@@ -19,14 +19,14 @@ import {
 } from '../../webapp_data_source/tb_http_client_testing';
 import {HparamsDataSource} from './hparams_data_source';
 import {
-  createHparamsExperimentResponse,
-  createHparamsListSessionGroupResponse,
-} from './testing';
-import {
   BackendHparamsValueType,
   DatasetType,
   DomainType,
   SessionGroup,
+  BackendHparamSpec,
+  BackendHparamsExperimentResponse,
+  BackendListSessionGroupResponse,
+  RunStatus,
 } from '../types';
 
 describe('HparamsDataSource Test', () => {
@@ -182,3 +182,193 @@ describe('HparamsDataSource Test', () => {
     });
   });
 });
+
+export function createHparamsListSessionGroupResponse(): BackendListSessionGroupResponse {
+  return {
+    sessionGroups: [
+      {
+        name: 'session_id_1',
+        hparams: {
+          hparams1: -100,
+          hparams2: 'bar',
+        },
+        sessions: [
+          {
+            endTimeSecs: 0,
+            metricValues: [
+              {
+                name: {
+                  group: '',
+                  tag: 'metrics1',
+                },
+                trainingStep: 1000,
+                value: 1,
+                wallTimeSecs: 0,
+              },
+            ],
+            modelUri: '',
+            monitorUrl: '',
+            name: 'run_name_1',
+            startTimeSecs: 0,
+            status: RunStatus.STATUS_SUCCESS,
+          },
+        ],
+      },
+      {
+        name: 'session_id_2',
+        hparams: {
+          hparams1: 100,
+          hparams2: 'foo',
+        },
+        sessions: [
+          {
+            endTimeSecs: 0,
+            metricValues: [
+              {
+                name: {
+                  group: 'train',
+                  tag: 'metrics1',
+                },
+                trainingStep: 2000,
+                value: 0.1,
+                wallTimeSecs: 0,
+              },
+              {
+                name: {
+                  group: 'test',
+                  tag: 'metrics1',
+                },
+                trainingStep: 5000,
+                value: 0.6,
+                wallTimeSecs: 0,
+              },
+            ],
+            modelUri: '',
+            monitorUrl: '',
+            name: 'run_name_2',
+            startTimeSecs: 0,
+            status: RunStatus.STATUS_SUCCESS,
+          },
+          {
+            endTimeSecs: 0,
+            metricValues: [
+              {
+                name: {
+                  group: 'train',
+                  tag: 'metrics1',
+                },
+                trainingStep: 10000,
+                value: 0.3,
+                wallTimeSecs: 0,
+              },
+              {
+                name: {
+                  group: 'train',
+                  tag: 'metrics2',
+                },
+                trainingStep: 10000,
+                value: 0,
+                wallTimeSecs: 0,
+              },
+            ],
+            modelUri: '',
+            monitorUrl: '',
+            name: 'run_name_2',
+            startTimeSecs: 0,
+            status: RunStatus.STATUS_RUNNING,
+          },
+        ],
+      },
+    ],
+    totalSize: 2,
+  };
+}
+
+export function createHparamsExperimentResponse(): BackendHparamsExperimentResponse {
+  return {
+    description: 'some description',
+    hparamInfos: [
+      {
+        description: 'describes hparams one',
+        displayName: 'hparams one',
+        name: 'hparams1',
+        type: BackendHparamsValueType.DATA_TYPE_STRING,
+        domainInterval: {minValue: -100, maxValue: 100},
+      },
+      {
+        description: 'describes hparams two',
+        displayName: 'hparams two',
+        name: 'hparams2',
+        type: BackendHparamsValueType.DATA_TYPE_BOOL,
+        domainDiscrete: ['foo', 'bar', 'baz'],
+      },
+    ],
+    metricInfos: [
+      {
+        name: {
+          group: '',
+          tag: 'metrics1',
+        },
+        displayName: 'Metrics One',
+        description: 'describe metrics one',
+        datasetType: DatasetType.DATASET_UNKNOWN,
+      },
+      {
+        name: {
+          group: 'group',
+          tag: 'metrics2',
+        },
+        displayName: 'Metrics Two',
+        description: 'describe metrics two',
+        datasetType: DatasetType.DATASET_TRAINING,
+      },
+    ],
+    name: 'experiment name',
+    timeCreatedSecs: 1337,
+    user: 'user name',
+  };
+}
+
+export function createHparamsExperimentNoDomainResponse(): BackendHparamsExperimentResponse {
+  return {
+    description: 'some description',
+    hparamInfos: [
+      {
+        description: 'describes hparams one',
+        displayName: 'hparams one',
+        name: 'hparams1',
+        type: BackendHparamsValueType.DATA_TYPE_STRING,
+      } as BackendHparamSpec,
+      {
+        description: 'describes hparams two',
+        displayName: 'hparams two',
+        name: 'hparams2',
+        type: BackendHparamsValueType.DATA_TYPE_BOOL,
+        domainDiscrete: ['foo', 'bar', 'baz'],
+      },
+    ],
+    metricInfos: [
+      {
+        name: {
+          group: '',
+          tag: 'metrics1',
+        },
+        displayName: 'Metrics One',
+        description: 'describe metrics one',
+        datasetType: DatasetType.DATASET_UNKNOWN,
+      },
+      {
+        name: {
+          group: 'group',
+          tag: 'metrics2',
+        },
+        displayName: 'Metrics Two',
+        description: 'describe metrics two',
+        datasetType: DatasetType.DATASET_TRAINING,
+      },
+    ],
+    name: 'experiment name',
+    timeCreatedSecs: 1337,
+    user: 'user name',
+  };
+}
