+++ tensorboard/webapp/runs/data_source/BUILD
@@ -39,7 +39,6 @@ tf_ng_module(
     deps = [
         ":backend_types",
         ":data_source",
-        "//tensorboard/webapp/hparams/_redux:testing",
         "@npm//@angular/core",
         "@npm//rxjs",
     ],
