+++ tensorboard/webapp/runs/data_source/runs_data_source_types.ts
@@ -63,6 +63,7 @@ export interface HparamSpec
   domain: Domain;
 }
 
+// TODO(rileyajones) merge these types by deleting this one.
 export interface MetricSpec extends backendTypes.MetricSpec {
   tag: string;
 }
